chore: refactor vaultsList component data
