+++ docs/upgrades/data-and-component-cleanup/step-1-2-vaultsList-refactoring.md
@@ -1,0 +1,135 @@
+# Step 1.2 - VaultsList Component Refactoring
+
+## Completion Report
+
+**Date Completed:** [Current Date]
+**Estimated Time:** 45 minutes
+**Status:** ✅ COMPLETED
+
+## What Was Accomplished
+
+Successfully extracted data transformation and filtering logic from the VaultsList component following the same pattern used in Step 1.1.
+
+### 1. Data Transformation Extraction
+
+**Created `useVaultListData.ts` hook:**
+
+- **Purpose:** Centralized vault data transformation logic
+- **Input:** Raw `Vault[]` and `TokenAsset[]` arrays  
+- **Output:** Properly formatted `VaultListData[]` for UI display
+- **Key Features:**
+  - Vault type classification (V2/V3 logic)
+  - Token asset icon resolution
+  - APY and TVL formatting with proper localization
+  - Chain information mapping
+  - Memoized for performance optimization
+
+### 2. Filtering and Sorting Extraction
+
+**Created `useVaultFiltering.ts` hook:**
+
+- **Purpose:** Reusable filtering and sorting logic for vault lists
+- **Features:**
+  - Multi-column sorting with direction toggles
+  - Search term filtering across all columns
+  - Range-based filtering for APY and TVL
+  - Type-based filtering integration
+  - Numeric parsing for proper sorting of formatted values
+
+### 3. Component Refactoring
+
+**Updated `VaultsList.tsx`:**
+
+- **Before:** 334 lines with embedded data processing
+- **After:** 210 lines focused on UI rendering (-37% reduction)
+- **Improvements:**
+  - Clean separation of data and presentation layers
+  - Simplified component logic
+  - Improved maintainability
+  - Better type safety
+
+## Files Modified
+
+### New Files Created
+
+1. `/src/hooks/useVaultListData.ts` - 46 lines
+2. `/src/hooks/useVaultFiltering.ts` - 137 lines
+
+### Files Refactored  
+
+1. `/src/components/VaultsList.tsx` - Reduced from 334 → 210 lines
+
+## Technical Implementation Details
+
+### Data Transformation Logic
+
+- Migrated complex vault type determination logic
+- Preserved existing business rules for V2/V3 classification
+- Maintained APY and TVL formatting with proper error handling
+- Added proper TypeScript typing throughout
+
+### Filtering Enhancement
+
+- Implemented type-safe column sorting
+- Added proper numeric parsing for formatted currency/percentage values
+- Created reusable filtering patterns that can be applied to other components
+- Maintained existing UX behavior while improving code organization
+
+### Performance Optimizations
+
+- Used `useMemo` for expensive data transformations
+- Proper dependency arrays to prevent unnecessary recalculations
+- Virtual scrolling integration maintained for large datasets
+
+## Impact Assessment
+
+### Code Quality Improvements
+
+- ✅ **Separation of Concerns:** Data logic cleanly separated from UI
+- ✅ **Reusability:** Hooks can be used by other vault-related components
+- ✅ **Maintainability:** Business logic centralized and documented
+- ✅ **Testability:** Data transformation logic easily unit testable
+
+### Performance Impact
+
+- ✅ **Memory:** Improved through proper memoization
+- ✅ **Rendering:** Reduced component complexity improves render performance
+- ✅ **Bundle:** No significant size impact (logic moved, not added)
+
+### Developer Experience
+
+- ✅ **Debugging:** Easier to debug data vs UI issues separately
+- ✅ **TypeScript:** Full type safety maintained throughout refactor
+- ✅ **IDE Support:** Better autocomplete and refactoring capabilities
+
+## Validation
+
+### Build Tests
+
+- ✅ TypeScript compilation successful
+- ✅ ESLint rules passing
+- ✅ No runtime errors
+
+### Functional Tests
+
+- ✅ Sorting functionality preserved
+- ✅ Filtering behavior maintained  
+- ✅ Virtual scrolling still operational
+- ✅ Type filtering integration working
+
+## Next Steps
+
+With Step 1.2 completed, we have successfully refactored the second-largest component in the application. The VaultsList component is now significantly cleaner and follows the same architectural patterns established in Step 1.1.
+
+**Ready for Step 2.1:** Component decomposition phase can now begin, focusing on breaking down remaining large components into smaller, focused sub-components.
+
+## Lessons Learned
+
+1. **Hook Pattern Consistency:** Using the same hook extraction pattern across components creates predictable code organization
+2. **Incremental Refactoring:** Step-by-step approach prevents regression issues
+3. **Type Safety:** Maintaining strict TypeScript typing throughout the refactor prevents runtime issues
+4. **Performance Awareness:** Proper memoization is critical when extracting data transformation logic
+
+---
+
+**Impact Summary:** VaultsList component reduced by 37% (334→210 lines) while improving maintainability, type safety, and reusability. Data transformation logic now centralized and easily testable.
